+++ extensions/jetbrains/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Jetbrains Changelog
 
+## [Toolbox V3 Support] - {PR_MERGE_DATE}
+- Added support for Toolbox V3
+
 ## [Bugfix] - 2025-04-14
 - Fix sorting projects which doesn't have a opened value
 - Fix crashing when `projects` field in `settings.json` is not present